GIT LFS

+++ datapickle.py
@@ -7,6 +7,6 @@ import pickle
 
 path = sys.argv[1]
 serial_path = os.path.splitext(path)[0] + '.pickle'
-arr = pd.read_csv(path, sep=",", header=None, index_col=None, dtype="float16")
+arr = pd.read_csv(path, sep=",", header=None, index_col=None, dtype="float32")
 print(arr.shape)
 pickle.dump(arr, open(serial_path, 'wb'))
